split student context
