+++ frontend/src/components/assessments/student-experience/QuestionNumbers.tsx
@@ -2,13 +2,16 @@ import React, { useContext } from "react";
 import { GridItem, SimpleGrid } from "@chakra-ui/react";
 
 import StudentContext from "../../../contexts/StudentContext";
+import WriteAssessmentContext from "../../../contexts/WriteAssessmentContext";
 import { questionStatus } from "../../../utils/StudentUtils";
 
 import QuestionNumber from "./QuestionNumber";
 
 const QuestionNumbers = (): React.ReactElement => {
-  const { test, answers, currentQuestionIndex, setCurrentQuestionIndex } =
-    useContext(StudentContext);
+  const { test } = useContext(StudentContext);
+  const { answers, currentQuestionIndex, setCurrentQuestionIndex } = useContext(
+    WriteAssessmentContext,
+  );
 
   return (
     <SimpleGrid columns={3} spacing={4}>
