+++ frontend/src/components/pages/student/StudentRouting.tsx
@@ -7,8 +7,6 @@ import type { TestResponse } from "../../../APIClients/types/TestClientTypes";
 import type { TestSessionSetupData } from "../../../APIClients/types/TestSessionClientTypes";
 import * as Routes from "../../../constants/Routes";
 import StudentContext from "../../../contexts/StudentContext";
-import type { Answers } from "../../../types/AnswerTypes";
-import { initializeAnswers } from "../../../utils/StudentUtils";
 import PrivateRoute from "../../auth/PrivateRoute";
 import ErrorState from "../../common/ErrorState";
 import LoadingState from "../../common/LoadingState";
@@ -28,10 +26,6 @@ const StudentRouting = (): React.ReactElement => {
     null,
   );
   const [className, setClassName] = useState("");
-  const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
-  const [answers, setAnswers] = useState<Answers[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isSubmitted, setIsSubmitted] = useState(false);
 
   useEffect(() => {
     if (state) {
@@ -49,7 +43,6 @@ const StudentRouting = (): React.ReactElement => {
     onCompleted: () => {
       if (data) {
         setTest(data.test);
-        setAnswers(initializeAnswers(data.test.questions));
       }
     },
   });
@@ -63,14 +56,6 @@ const StudentRouting = (): React.ReactElement => {
         setTestSession,
         className,
         setClassName,
-        currentQuestionIndex,
-        setCurrentQuestionIndex,
-        answers,
-        setAnswers,
-        isLoading,
-        setIsLoading,
-        isSubmitted,
-        setIsSubmitted,
       }}
     >
       {loading && <LoadingState fullPage />}
